+++ R/app_ui.R
@@ -47,18 +47,11 @@ TutorialUI <- function() {
 AttributionResultsUI <- function() {
   fluidPage(
     br(),
-    fluidRow(splitLayout(cellWidths = c("25%", "25%"),
+    fluidRow(splitLayout(cellWidths = c("25%", "25%", "17.5%", "20%"),
                          actionButton(inputId = "analyzeMoreSample", 
-                          label = "Analyze another sample",
-                          style = "color: #fff; background-color: #337ab7;
+                                      label = "Do another signature attribution",
+                                      style = "color: #fff; background-color: #337ab7;
                              border-color: #2e6da4;"),
-                         bookmarkButton(label = "Save the session", 
-                                        title = paste0("The session will be saved ", 
-                                                       "on the server for seven days"),
-                                        style = "color: #fff; background-color: #337ab7;
-                             border-color: #2e6da4;"))),
-    br(),
-    fluidRow(splitLayout(cellWidths = c("25%", "17.5%", "20%"),
                          downloadButton(outputId = "downloadExposureTable", 
                                         label = "Download exposure counts"),
                          downloadButton(outputId = "downloadPdf",